+++ apps/designer/app/designer/shared/assets/update-asset-containers.test.ts
@@ -6,7 +6,7 @@ import {
   DeletingAssetContainer,
   AssetContainer,
 } from "./types";
-import { updateStateAssets } from "./update-state-assets";
+import { updateStateAssets } from "./update-asset-containers";
 
 const getAssetId = (
   assetContainer: AssetContainer | DeletingAssetContainer
